Handle several GHC versions
Add several GHC interface files in HImport scan and several snapshot.yaml
to allow the user to use its favourite ghc version using --config=ghc_X_X_X.
The three versions of GHC in nixpks 22.05 (version 8.10.7, 9,.0.2 and 9.2.2)
are supported now.
For version 9.2.2, due to a bug in version 3.6.3.0 of cabal, one has to use
a custom version of cabal.